Prototype "cred bounties" via manual weights
This is a small experiment in having something like [Cred Bounties][1]
before we've actually implemented the [Initiatives Plugin][2].

Basically, we're giving a higher weight to a few initiatives:
- Produce the SourceCred Podcast: Shipped the podcast!
- Discourse Reference Detection: Recently completed.
- Champions & Heroes: An exciting contribution. :)
- Initiatives Plugin: Nice progress so far.

The weight choices are moderate, so it doesn't have a _huge_ effect on
cred--we'll hold off on big effects for after the initiatives plugin and
cred bounties both land. But this is a nice way to recognize some
exciting progress so far.

[1]: https://discourse.sourcecred.io/t/enable-cred-bounties/257
[2]: https://discourse.sourcecred.io/t/write-the-initiatives-plugin/269