convert weights to 0.2.0 format
This update is needed as of sourcecred/sourcecred#1558. I didn't make
any changes to the weights themselves. As discussed in the review there,
the weights semantics changed in some edge cases; happily, none of those
edge cases were present in our weights.

Test plan: I've re-computed cred using the new weights (built against
sourcecred master) and verified that the weights are set correctly in
the UI.