Commit dist/api.js to sourcecred/cred
Conceputally speaking, it's not clear that the API should be bundled
with the cred. In practice, it's extremely convenient for setting up
cred-analysis notebooks, as it means the notebook can load the cred data
and a consonant API version from the same repository.

Test plan: Need to commit this to master and push before we'll see
whether loading it from GitHub pages works.