Switch to like-minted-cred (#16)
This changes the weights so that on Discourse, we only mint cred based
on likes, rather than to raw activity. I've set the like-weight to 16,
which is the same as a GitHub pull request. This is a somewhat arbitrary
decision and likely to be re-evaluated in the future.

See extensive discussion around this change at [this Discourse post][1].

Test plan: Review, and I'll compile a full score update as a followon.

[1]: https://discourse.sourcecred.io/t/minting-discourse-cred-on-likes-not-posts/603