Preview: Add a first batch of initiatives to the project
## This is a preview

This is a preview, as it's in need of another weighting round.
Still missing from this, but identified at the team call:
- CredCon: video interviews (need a user account for zak)
- CredCon: CredRank prototype (have not decided fibonacci weight)
- CredCon: Discord prototype (have not decided fibonacci weight)
- Evan's explainer (have not decided fibonacci weight)

Also great improvements to make from here:
- Adding contributions and linking that to contributors

Good but not as critical:
- Making the timestamps more accurate


## Assumptions to validate


### Fibonacci to absolute weights: 200x

The major point to debate is the weights we should set. In the team call
we've decided on fibonacci scale weights, which I've included in these
files as `_fibonacciWeight`.

This PR includes a ballpark suggestion of a multiplier to use.
The intention is to "inflate cred" by weighing initiatives more heavily
than plugin-imported contributions so far. While keeping in mind that
more initiatives will follow and inflate cred further. And adding more
contribution edges will "dilute" the share currently flowing to champions.

A multiplier which I think is in this ballpark is: 200x
(e.g. a fibonacci weight of 2 = a node weight of 400)


### Dependency structure

For the activities as part of CredCon (like hackathon, party, talks), I
found they had very similar dependencies on logistical initiatives and
CredCon itself. I've generalized this by depending on CredCon, rather than
directly depending on logistics.

Example: CredCon party, depends on CredCon, depends on car+food+lodging

This will make the life of historians easier too, as adding more logicstics
(maybe invites or scheduling) would only need to be set as a CredCon
dependency, rather than for all activities as well.


### Example references / contributions

For the main CredCon initiative, I've added several clearly related forum
topics. These are both as an example, and to make sure this part of the
graph is not exclusively connected through identities.


### Completion status

Other than both Partner initiatives, the CredCon initiatives are set as
completed.


## Technical aspects to validate

The main point which needs careful manual validation, is to inspect the
graph. For example using the "legacy explorer" to confirm the initiative
nodes are added and have edges to *all* the URLs added.

That means:
- Champions: Identities / GitHub users / Discourse users.
- References / Contributions: Forum topics (in CredCon).
- Dependencies between Initiatives.

Note: it's expected for links to an initiative to give a 404 not found,
until the PR is merged to master. However they should be listed in the
explorer.