Add an initiative for our ongoing docs push
We have a lot of good work happening in sourcecred/docs, and I felt it
would be a good opportunity for more experimental usage of our
initiatives system. I went through the history of the repo and added
major pull requests as contributions; my heuristic was that included
pull requests should either:
- contribute material new content to the documentation, or
- make material technical progress towards hosting and serving the docs

I did not include every pull request as some are duplicates (e.g. the
same content was placed in a few pull requests) and some didn't meet my
materiality threshold. When we add the ability to set
'contribution weights' we will be able to do this in a more fine-grained
way.

I also added several forum posts and issues as 'references'.

If you feel that some contributions should/should not be included, feel
free to say so, focusing on the value of the contribution in question.
Please keep in mind that [initiatives are still experimental][1].

I set the cred weight as 200 when incomplete, and 1000 when complete.
This feels roughly in line with our current cred totals (we may revise
this later).

[1]: https://discourse.sourcecred.io/t/rolling-out-initiatives/698

Test plan: Initiative loads without parse errors, and the resultant cred
is different from cred before (people who participated have slightly
more cred). It's not a huge effect because I set this initiative as
worth 200 cred while incomplete.