Add first batch of CredCon and Partner initiatives (#27)
* Preview: Add a first batch of initiatives to the project

## This is a first iteration

Rather than letting this stall for too long, let's improve on this with follow ups.
Have a look at #29 for info on how to go from here.

This is a first iteration, as it's in need of another weighting round.
Still missing from this, but identified at the team call:
- CredCon: video interviews (need a user account for zak)
- CredCon: CredRank prototype (have not decided fibonacci weight)
- CredCon: Discord prototype (have not decided fibonacci weight)
- Evan's explainer (have not decided fibonacci weight)

Also great improvements to make from here:
- Adding contributions and linking that to contributors (see #30)

Good but not as critical:
- Making the timestamps more accurate

## Fibonacci to absolute weights: 100x first, 200x later

The 200x multiplier seems in the right ballpark, however we want to
avoid champions see Cred "dilute" because we'll add more contributions.

Start with 100x now, and move to 200x as we have a more complete coverage.
Tracking this in issue #31.

## Other assumptions made

### Dependency structure

For the activities as part of CredCon (like hackathon, party, talks), I
found they had very similar dependencies on logistical initiatives and
CredCon itself. I've generalized this by depending on CredCon, rather than
directly depending on logistics.

Example: CredCon party, depends on CredCon, depends on car+food+lodging

This will make the life of historians easier too, as adding more logicstics
(maybe invites or scheduling) would only need to be set as a CredCon
dependency, rather than for all activities as well.

### Example references / contributions

For the main CredCon initiative, I've added several clearly related forum
topics. These are both as an example, and to make sure this part of the
graph is not exclusively connected through identities.

### Completion status

Other than both Partner initiatives, the CredCon initiatives are set as
completed.

## Technical aspects to validate

The main point which needs careful manual validation, is to inspect the
graph. For example using the "legacy explorer" to confirm the initiative
nodes are added and have edges to *all* the URLs added.

That means:
- Champions: Identities / GitHub users / Discourse users.
- References / Contributions: Forum topics (in CredCon).
- Dependencies between Initiatives.

Note: it's expected for links to an initiative to give a 404 not found,
until the PR is merged to master. However they should be listed in the
explorer.