Update update.sh to include CNAME
Now that we've verified that the CNAME fix from #46 works, let's
canonicalize it in the update script so that we won't need to maintain
it by hand.

I ran update.sh and committed the CNAME file generated inside docs, its
the same modulo trailing newlines.

Test plan: Run update.sh, and verify that the CNAME file is no longer
removed or modified.