Update to include more recent contributions
In combo with @YalorMewn and @hammadj, added some more contributions
over the past month.

I went over the weights on my own and tried to make things multiples of
16, because we mint 16 Cred per PR/Discourse Like/SourceCred emoji. I
erred on the side of being conservative, since we'll flow a lot more
Cred once we've actually delivered a working editor.

Test plan: Re-compute Cred, check for "user not found" warnings.