add commit step to grain publish workflow
add step to pubilsh ledger.json back to the repo after recalculating
grain. This will publish with the default "Github Action" as the
commiter but we can customize this if we want

one thing to consider: we need to esnure the token to execute this
workflow has push access to the repo. If the non-default token doesn't
have push access we'll need to set this token in checkout step

test plan: this can be tested locally (with some tweaking0 using
https://github.com/nektos/act. It is recommended that an existing test
branch be specified on the commit action in the `branch` yaml options