test distribution workflow
this is a testing ground for the grain distribution github workflow.
nothing will be published to the gh-pages sites. Any changes will be
pushed back to this branch. the ledger state has been rolled back to
ensure distributions execute and the ledger is appended to  state that
can be pushed.

test-plan: observe the github action to see if the updated ledger is
pushed back to this branch from the workflow itself. Ensure new
distritbutions are recorded in the new ledger and the frequency of those
distributions matches master