Grain sale to decentralion
This records a sale of Grain by the SourceCred treasury to @decentralion
(myself). Usually, the treasury sells Grain to Protocol Labs, but we
needed liquidity for an urgent Grain redemption.

Since my account is currently de-activated pending implementation of a
salary cap[1], I re-activated it, then did the transfer, then
de-activated.

[1]: https://discourse.sourcecred.io/t/adding-a-grain-salary-cap/925/16