Stop minting Cred for SourceCred
This commit makes us stop minting Cred for SourceCred (and removes this
retro-actively). We set this up so that SC treasury would have a source
of Grain getting minted, however we're going to switch to a new approach
where we bulk-mint Grain directly to SourceCred, rather than having SC
"skim" a portion of the weekly distributions.